refact: image -> file로 통합
