+++ src/main/java/umc/cockple/demo/domain/file/exception/S3Exception.java
@@ -1,0 +1,9 @@
+package umc.cockple.demo.domain.file.exception;
+
+import umc.cockple.demo.global.exception.GeneralException;
+import umc.cockple.demo.global.response.code.BaseErrorCode;
+
+public class S3Exception extends GeneralException {
+
+    public S3Exception(BaseErrorCode code) { super(code); }
+}
