+++ src/main/java/umc/cockple/demo/domain/file/service/FileService.java
@@ -1,0 +1,110 @@
+package umc.cockple.demo.domain.file.service;
+
+import com.google.cloud.storage.*;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
+import umc.cockple.demo.domain.file.dto.FileUploadDTO;
+import umc.cockple.demo.domain.file.exception.S3ErrorCode;
+import umc.cockple.demo.domain.file.exception.S3Exception;
+import umc.cockple.demo.global.enums.DomainType;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class FileService {
+
+    @Value("${gcs.bucket}")
+    private String bucket;
+
+    private final Storage storage;
+
+    public FileUploadDTO.Response uploadFile(MultipartFile file, DomainType domainType) {
+        if (file == null || file.isEmpty()) {
+            return null;
+        }
+
+        log.info("[파일 업로드 시작]");
+
+        String originalFileName = file.getOriginalFilename();
+        String key = getFileKey(file, domainType);
+        String fileUrl = uploadToGcs(file, key);
+
+        log.info("[파일 업로드 완료]");
+        return FileUploadDTO.Response.builder()
+                .fileKey(key)
+                .fileUrl(fileUrl)
+                .originalFileName(originalFileName)
+                .fileSize(file.getSize())
+                .fileType(file.getContentType())
+                .build();
+    }
+
+    public List<FileUploadDTO.Response> uploadFiles(List<MultipartFile> files, DomainType domainType) {
+        if (files == null || files.isEmpty()) {
+            return List.of();
+        }
+
+        return files.stream()
+                .map(img -> uploadFile(img, domainType))
+                .collect(Collectors.toList());
+    }
+
+    public void delete(String fileKey) {
+        try {
+            storage.delete(BlobId.of(bucket, fileKey));
+            log.info("[GCS 삭제 성공] {}", fileKey);
+        } catch (Exception e) {
+            log.error("[GCS 삭제 실패] {}", e.getMessage());
+            throw new S3Exception(S3ErrorCode.IMAGE_DELETE_EXCEPTION);
+        }
+    }
+
+    private String uploadToGcs(MultipartFile file, String key) {
+        try {
+            BlobInfo blobInfo = BlobInfo.newBuilder(bucket, key)
+                    .setContentType(file.getContentType())
+                    .build();
+            storage.create(blobInfo, file.getBytes());
+            return String.format("https://storage.googleapis.com/%s/%s", bucket, key);
+        } catch (IOException e) {
+            log.error("[GCS 업로드 실패 - IO 예외] {}", e.getMessage());
+            throw new S3Exception(S3ErrorCode.FILE_UPLOAD_IO_EXCEPTION);
+        } catch (StorageException e) {
+            log.error("[GCS 업로드 실패 - Storage 예외] {}", e.getMessage());
+            throw new S3Exception(S3ErrorCode.FILE_UPLOAD_AMAZON_EXCEPTION);
+        }
+    }
+
+    public String getFileKey(MultipartFile file, DomainType domainType) {
+        if (file == null || file.isEmpty()) {
+            return null;
+        }
+
+        String originalFilename = file.getOriginalFilename();
+        String extension = StringUtils.getFilenameExtension(originalFilename);
+        String uuid = UUID.randomUUID().toString();
+
+        return domainType.getDirectory() + "/" + uuid + "." + extension;
+    }
+
+    public String getUrlFromKey(String key) {
+        return String.format("https://storage.googleapis.com/%s/%s", bucket, key);
+    }
+
+    public Blob downloadFile(String fileKey) {
+        Blob blob = storage.get(BlobId.of(bucket, fileKey));
+        if (blob == null) {
+            throw new S3Exception(S3ErrorCode.IMAGE_DELETE_EXCEPTION);
+        }
+        return blob;
+    }
+}
