+++ src/test/java/umc/cockple/demo/domain/chat/service/ChatProcessorTest.java
@@ -14,7 +14,7 @@ import umc.cockple.demo.domain.chat.domain.ChatMessageImg;
 import umc.cockple.demo.domain.chat.domain.ChatRoom;
 import umc.cockple.demo.domain.chat.dto.ChatCommonDTO;
 import umc.cockple.demo.domain.chat.enums.MessageType;
-import umc.cockple.demo.domain.image.service.ImageService;
+import umc.cockple.demo.domain.file.service.FileService;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.member.domain.ProfileImg;
 import umc.cockple.demo.global.enums.Gender;
@@ -35,7 +35,7 @@ import static org.mockito.Mockito.verify;
 class ChatProcessorTest {
 
     @Mock
-    private ImageService imageService;
+    private FileService fileService;
 
     private ChatConverter chatConverter;
     private ChatProcessor chatProcessor;
@@ -46,7 +46,7 @@ class ChatProcessorTest {
     @BeforeEach
     void setUp() {
         chatConverter = new ChatConverter();
-        chatProcessor = new ChatProcessor(imageService, chatConverter);
+        chatProcessor = new ChatProcessor(fileService, chatConverter);
 
         sender = MemberFixture.createMemberWithName("홍길동", "길동", Gender.MALE, Level.A, 1001L);
         ReflectionTestUtils.setField(sender, "id", 10L);
@@ -69,7 +69,7 @@ class ChatProcessorTest {
             String result = chatProcessor.generateProfileImageUrl(null);
 
             assertThat(result).isNull();
-            verify(imageService, never()).getUrlFromKey(null);
+            verify(fileService, never()).getUrlFromKey(null);
         }
 
         @Test
@@ -82,7 +82,7 @@ class ChatProcessorTest {
             String result = chatProcessor.generateProfileImageUrl(profileImg);
 
             assertThat(result).isNull();
-            verify(imageService, never()).getUrlFromKey(null);
+            verify(fileService, never()).getUrlFromKey(null);
         }
 
         @Test
@@ -95,7 +95,7 @@ class ChatProcessorTest {
             String result = chatProcessor.generateProfileImageUrl(profileImg);
 
             assertThat(result).isNull();
-            verify(imageService, never()).getUrlFromKey("   ");
+            verify(fileService, never()).getUrlFromKey("   ");
         }
 
         @Test
@@ -105,13 +105,13 @@ class ChatProcessorTest {
                     .imgKey("profile/key123.jpg")
                     .build();
 
-            given(imageService.getUrlFromKey("profile/key123.jpg"))
+            given(fileService.getUrlFromKey("profile/key123.jpg"))
                     .willReturn("https://cdn.example.com/profile/key123.jpg");
 
             String result = chatProcessor.generateProfileImageUrl(profileImg);
 
             assertThat(result).isEqualTo("https://cdn.example.com/profile/key123.jpg");
-            verify(imageService).getUrlFromKey("profile/key123.jpg");
+            verify(fileService).getUrlFromKey("profile/key123.jpg");
         }
     }
 
@@ -143,7 +143,7 @@ class ChatProcessorTest {
             String result = chatProcessor.generateImageUrl(img);
 
             assertThat(result).isNull();
-            verify(imageService, never()).getUrlFromKey(null);
+            verify(fileService, never()).getUrlFromKey(null);
         }
 
         @Test
@@ -160,7 +160,7 @@ class ChatProcessorTest {
             String result = chatProcessor.generateImageUrl(img);
 
             assertThat(result).isNull();
-            verify(imageService, never()).getUrlFromKey("  ");
+            verify(fileService, never()).getUrlFromKey("  ");
         }
 
         @Test
@@ -174,13 +174,13 @@ class ChatProcessorTest {
                     .fileType("image/jpeg")
                     .build();
 
-            given(imageService.getUrlFromKey("chat/img456.jpg"))
+            given(fileService.getUrlFromKey("chat/img456.jpg"))
                     .willReturn("https://cdn.example.com/chat/img456.jpg");
 
             String result = chatProcessor.generateImageUrl(img);
 
             assertThat(result).isEqualTo("https://cdn.example.com/chat/img456.jpg");
-            verify(imageService).getUrlFromKey("chat/img456.jpg");
+            verify(fileService).getUrlFromKey("chat/img456.jpg");
         }
     }
 
@@ -277,7 +277,7 @@ class ChatProcessorTest {
             List<ChatCommonDTO.MessageInfo> result = chatProcessor.processMessages(sender.getId(), List.of(message));
 
             assertThat(result.get(0).senderProfileImageUrl()).isNull();
-            verify(imageService, never()).getUrlFromKey(null);
+            verify(fileService, never()).getUrlFromKey(null);
         }
 
         @Test
@@ -336,9 +336,9 @@ class ChatProcessorTest {
             // ChatMessage의 chatMessageImgs에 역순으로 세팅
             ReflectionTestUtils.setField(message, "chatMessageImgs", List.of(img3, img1, img2));
 
-            given(imageService.getUrlFromKey("img/first.jpg")).willReturn("https://cdn.example.com/first.jpg");
-            given(imageService.getUrlFromKey("img/second.jpg")).willReturn("https://cdn.example.com/second.jpg");
-            given(imageService.getUrlFromKey("img/third.jpg")).willReturn("https://cdn.example.com/third.jpg");
+            given(fileService.getUrlFromKey("img/first.jpg")).willReturn("https://cdn.example.com/first.jpg");
+            given(fileService.getUrlFromKey("img/second.jpg")).willReturn("https://cdn.example.com/second.jpg");
+            given(fileService.getUrlFromKey("img/third.jpg")).willReturn("https://cdn.example.com/third.jpg");
 
             List<ChatCommonDTO.MessageInfo> result = chatProcessor.processMessages(sender.getId(), List.of(message));
