+++ src/test/java/umc/cockple/demo/domain/exercise/service/ExerciseLifecycleServiceTest.java
@@ -17,7 +17,7 @@ import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
 import umc.cockple.demo.domain.exercise.exception.ExerciseException;
 import umc.cockple.demo.domain.exercise.repository.ExerciseRepository;
 import umc.cockple.demo.domain.exercise.service.command.internal.ExerciseLifecycleService;
-import umc.cockple.demo.domain.image.service.ImageService;
+import umc.cockple.demo.domain.file.service.FileService;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.member.repository.MemberExerciseRepository;
 import umc.cockple.demo.domain.member.repository.MemberPartyRepository;
@@ -49,7 +49,7 @@ class ExerciseLifecycleServiceTest {
     @Mock private PartyRepository partyRepository;
     @Mock private MemberPartyRepository memberPartyRepository;
     @Mock private MemberExerciseRepository memberExerciseRepository;
-    @Mock private ImageService imageService;
+    @Mock private FileService fileService;
 
     private ExerciseLifecycleService exerciseLifecycleService;
 
@@ -59,7 +59,7 @@ class ExerciseLifecycleServiceTest {
     @BeforeEach
     void setUp() {
         ExerciseValidator exerciseValidator = new ExerciseValidator(memberPartyRepository, memberExerciseRepository);
-        ExerciseConverter exerciseConverter = new ExerciseConverter(imageService);
+        ExerciseConverter exerciseConverter = new ExerciseConverter(fileService);
         exerciseLifecycleService = new ExerciseLifecycleService(
                 exerciseRepository, partyRepository, exerciseValidator, exerciseConverter);
